backend/pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["samrqs <samarassk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.4"
django-admin = "^2.0.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
djangorestframework = "^3.15.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
